feature: adiciona as novas controller referentes ao usuário
